Add file signature support to rpmsign command
This patch extends the rpmsign tool to sign package files. It defines a new
rpmsign option called "signfiles".

rpm --addsign [--signfiles] PACKAGE

Signfiles signs all the file digests included in the package and stores
the signatures in the package header. The file signing key, used to sign
the file digests, can be provided one the command line with --fskpath or
in a macro file with %_file_signing_key. After including file signatures,
the package is signed normally.

The package needs to be built with SHA-1 or SHA-2 digests before package
files are signed, this prerequisite is noted in rpmsign man page.