Fix signfiles issue with shrinking sigh
When file signatures are added to a package the signature digests are
replaced. Sometimes the new signature header is smaller than the
original. In this case don't insert the new signature. Instead rewrite
the signature header and rest of package.

This fixes a bug when adding file signatures to packages built with
rpm-4.12.0 (or older version). When trying to install the signed package,
or check its signatures, there is an error 'headerRead failed: hdr magic:
BAD'.